pyproject.toml
[project]
name = "cdg-automation-framework"
version = "1.1.0"
description = "Scalable API test automation framework"
requires-python = ">=3.11"
readme = "README.md"
authors = [{ name = "Dmitriy Tikhonov" }]
license = { file = "LICENSE" }
dependencies = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
    "allure-pytest>=2.13.5",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.35",
    "psycopg2-binary>=2.9.9",
    "mimesis>=17.0.0",
    "pika>=1.3.2",
]

[build-system]
requires = ["setuptools>=70.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.2",
    "pre-commit>=3.8.0",
]

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "SIM"]

[tool.ruff.format]
quote-style = "double"
